add: verification of web3 user email on server (#59)
